backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
